# lib/api-zod/src/generated/types/clothingItemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItemInputCategory } from './clothingItemInputCategory';

export interface ClothingItemInput {
  /** @minLength 1 */
  name: string;
  category: ClothingItemInputCategory;
  imageObjectPath?: string;
  color?: string;
  brand?: string;
  notes?: string;
  isFavorite?: boolean;
}