# lib/api-zod/src/generated/types/generateOutfitOptions.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateOutfitOptions {
  excludeCategories?: string[];
}