# lib/api-zod/src/generated/types/generatedOutfit.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { ClothingItem } from './clothingItem';

export interface GeneratedOutfit {
  items: ClothingItem[];
}