# lib/api-zod/src/generated/types/outfitInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface OutfitInput {
  /** @minLength 1 */
  name: string;
  notes?: string;
  /** @minItems 1 */
  itemIds: number[];
}