# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from './uploadUrlRequest';

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}