# lib/api-zod/src/generated/types/wardrobeStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Outfit Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryStat } from './categoryStat';

export interface WardrobeStats {
  total: number;
  byCategory: CategoryStat[];
  favorites: number;
  outfitsGenerated: number;
}